+++ sphinx/themes/basic/static/searchindex.js
@@ -1,0 +1,4 @@
+req = new XMLHttpRequest();
+req.open("GET", "searchindex.json", false);
+req.send(null);
+Search.setIndex(req.responseText);